% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_P_estimates.R
\name{get_global_P_estimates}
\alias{get_global_P_estimates}
\title{Get median, 95% and 80% credible intervals for posterior samples of P from JAGS model}
\usage{
get_global_P_estimates(
  subnat_index_table,
  method_index_table,
  sector_type,
  year_index_table,
  Psamps
)
}
\arguments{
\item{subnat_index_table}{Dataframe with subnational district indexing applied. Used to match estimates to data.}

\item{method_index_table}{Dataframe with method indexing applied. Used to match estimates to data.}

\item{sector_type}{String. Name of sector you are interested in. One of either ("Public", "Commercial medical", "Other")}

\item{year_index_table}{Dataframe with time indexing applied. Used to match estimates to data.}

\item{Psamps}{posterior samples of P from JAGS model}
}
\value{
Dataframe of labelled posterior samples with median, 95% and 80% credible intervals estimates.
}
\description{
Get median, 95% and 80% credible intervals for posterior samples of P from JAGS model
}
\examples{
get_global_P_estimates(subnat_index_table, method_index_table, sector_type, year_index_table, Psamps)
}
