% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_P_samps.R
\name{get_global_P_samps}
\alias{get_global_P_samps}
\title{Get posterior samples of P from r and z variables of JAGS model}
\usage{
get_global_P_samps(main_path)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}
}
\value{
Saved samples for public, commercial medical and other supply shares.
}
\description{
Get posterior samples of P from r and z variables of JAGS model
}
