% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IPUMSdata.R
\name{get_ipumsdata}
\alias{get_ipumsdata}
\title{Get IPUMS data}
\usage{
get_ipumsdata(mycountry)
}
\arguments{
\item{mycountry}{The country of interest for your local mcmsupply estimation. A list of possible countries available found in data/mycountries.rda.}
}
\value{
the input data for your country of interest, used as an input to the mcmsupply model
}
\description{
Get IPUMS data
}
\examples{
mydata <- get_ipumsdata("Nepal")
}
