% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_P_median_quantile.R,
%   R/get_P_point_estimates.R
\name{get_P_point_estimates}
\alias{get_P_point_estimates}
\title{Get median, 95% and 80% credible intervals for posterior samples of P from JAGS model}
\usage{
get_P_point_estimates(
  main_path,
  pkg_data,
  local = FALSE,
  spatial = FALSE,
  mycountry = NULL
)

get_P_point_estimates(
  main_path,
  pkg_data,
  local = FALSE,
  spatial = FALSE,
  mycountry = NULL
)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}

\item{pkg_data}{Output of the `mcmsupplylocal::get_subnational_modelinputs()` function.}

\item{local}{TRUE/FALSE. Default is FALSE. local=FALSE retrieves the data for all subnational provinces across all countries. local=TRUE retrieves data for only one country.}

\item{spatial}{TRUE/FALSE. Default is FALSE. spatial=FALSE retrieves the data for all subnational provinces across all countries without GPS information. spatial=TRUE retrieves for data for countries with GPS information as well as FP source data.}

\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}

\item{subnat_index_table}{Dataframe with subnational district indexing applied. Used to match estimates to data.}

\item{method_index_table}{Dataframe with method indexing applied. Used to match estimates to data.}

\item{sector_type}{String. Name of sector you are interested in. One of either ("Public", "Commercial medical", "Other")}

\item{year_index_table}{Dataframe with time indexing applied. Used to match estimates to data.}

\item{Psamps}{posterior samples of P from JAGS model}
}
\value{
Dataframe of labelled posterior samples with median, 95% and 80% credible intervals estimates.

returns the point estimates for the jags model object
}
\description{
Get median, 95% and 80% credible intervals for posterior samples of P from JAGS model

Get the r and z ratio point estimates from the separate chains of the JAGS model runs
R and Z are the intermediate parameters that are used to estimates the final proportions. See the model file for context.
}
\examples{
get_P_point_estimates(subnat_index_table, method_index_table, sector_type, year_index_table, Psamps)
}
