% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_local_results.R
\name{plot_point_estimates}
\alias{plot_point_estimates}
\title{Plot median, 95% and 80% credible intervals for posterior samples of P from local JAGS model with the relevant survey data}
\usage{
plot_point_estimates(P_samps, pkg_data, vis_path)
}
\arguments{
\item{pkg_data}{Output of the `mcmsupplylocal::get_subnational_modelinputs()` function.}

\item{vis_path}{String. Path where your visualisations will be saved to.}

\item{Psamps}{Posterior samples of P for one sector from JAGS model}
}
\value{
Data frame of labelled posterior samples with median, 95% and 80% credible intervals estimates.
}
\description{
Plot median, 95% and 80% credible intervals for posterior samples of P from local JAGS model with the relevant survey data
}
