% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_subnational_jags_model.R
\name{run_subnational_jags_model}
\alias{run_subnational_jags_model}
\title{Run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model}
\usage{
run_subnational_jags_model(
  jagsdata,
  jagsparams,
  local = FALSE,
  spatial = FALSE,
  mycountry
)
}
\arguments{
\item{jagsdata}{The inputs for the JAGS model}

\item{jagsparams}{The parameters of the JAGS model you wish to review}

\item{local}{TRUE/FALSE. Default is FALSE. local=FALSE retrieves the data for all subnational provinces across all countries. local=TRUE retrieves data for only one country.}

\item{spatial}{TRUE/FALSE. Default is FALSE. spatial=FALSE retrieves the data for all subnational provinces across all countries without GPS information. spatial=TRUE retrieves for data for countries with GPS information as well as FP source data.}

\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}
}
\value{
returns the jags model object
importFrom("stats", "cor", "filter", "lag")
}
\description{
Run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model
}
